using BigDecimal computations
git-svn-id: https://svn.win.tue.nl/repos/prom/Packages/StochasticAwareConformanceChecking/Trunk@40169 a68f48d4-8d00-424f-b812-3f22a534a02a

+++ src/org/processmining/stochasticawareconformancechecking/plugins/EntropyPlugin.java
@@ -1,5 +1,7 @@
 package org.processmining.stochasticawareconformancechecking.plugins;
 
+import java.math.BigDecimal;
+
 import org.processmining.contexts.uitopia.annotations.UITopiaVariant;
 import org.processmining.framework.plugin.PluginContext;
 import org.processmining.framework.plugin.annotations.Plugin;
@@ -19,7 +21,7 @@ public class EntropyPlugin {
 	@PluginVariant(variantLabel = "Compute entropy of sdfa, dialog", requiredParameterLabels = { 0 })
 	public HTMLToString mineGuiProcessTree(final PluginContext context,
 			StochasticDeterministicFiniteAutomaton automaton) throws UnsupportedAutomatonException {
-		final double entropy = Entropy.entropy(automaton);
+		final BigDecimal entropy = Entropy.entropy(automaton);
 		return new HTMLToString() {
 			public String toHTMLString(boolean includeHTMLTags) {
 				return "entropy: " + entropy;
